+++ src/js/policies/firefox.js
@@ -231,6 +231,25 @@ const policies = {
             value : 'false'
           }
         ]
+      },
+      {
+        name : 'MoreFromMozilla',
+        label : browser.i18n.getMessage('policy_description_UserMessaging_MoreFromMozilla'),
+        type : 'enum',
+        options : [
+          {
+            label : browser.i18n.getMessage('enum_value_no_preference'),
+            value : null
+          },
+          {
+            label : browser.i18n.getMessage('enum_value_yes'),
+            value : 'true'
+          },
+          {
+            label : browser.i18n.getMessage('enum_value_no'),
+            value : 'false'
+          }
+        ]
       }
     ]
   },